chore: rename new select
